refactor: Position class(객체) 분리
