+++ src/test/java/racingTest/CarTest.java
@@ -3,11 +3,7 @@ package racingTest;
 import org.junit.jupiter.api.Test;
 import racing.model.Car;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 
 public class CarTest {
@@ -27,13 +23,5 @@ public class CarTest {
         }).isInstanceOf(NullPointerException.class)
         .hasMessage("차이름은 공백 또는 빈 문자열일 수 없습니다.");
     }
-
-    @Test
-    void 현재_포지션_만큼_연속된_대쉬_문자열반환(){
-        Car car = new Car("boss");
-        String currentPositionByDash = car.getPositionByDash(5);
-        assertThat(currentPositionByDash).isEqualTo("-----");
-    }
-
 
 }
